+++ pyathena/sqlalchemy/util.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+
+
+class HashableDict(dict):  # type: ignore
+    def __hash__(self):
+        return hash(tuple(sorted(self.items())))
