+++ pyproject.toml
@@ -75,10 +75,10 @@ arrow = ["pyarrow"]
 fastparquet = ["fastparquet"]
 
 [tool.poetry.plugins."sqlalchemy.dialects"]
-"awsathena" = "pyathena.sqlalchemy_athena:AthenaDialect"
-"awsathena.rest" = "pyathena.sqlalchemy_athena:AthenaRestDialect"
-"awsathena.pandas" = "pyathena.sqlalchemy_athena:AthenaPandasDialect"
-"awsathena.arrow" = "pyathena.sqlalchemy_athena:AthenaArrowDialect"
+"awsathena" = "pyathena.sqlalchemy.base:AthenaDialect"
+"awsathena.rest" = "pyathena.sqlalchemy.rest:AthenaRestDialect"
+"awsathena.pandas" = "pyathena.sqlalchemy.pandas:AthenaPandasDialect"
+"awsathena.arrow" = "pyathena.sqlalchemy.arrow:AthenaArrowDialect"
 
 [tool.pytest.ini_options]
 norecursedirs = ["benchmarks", ".poetry", ".tox"]
